Force $TERM to be "alacritty" when using alacritty.
Note: While the comment in the config says, it is attempting to check
whether a terminfo for alacritty exists on the system and then adjust
the $TERM based on that, in reality that is not working on windows.
Granted, this is really hard, given that the launched shell may be
powershell or cmd which have no notion of terminfo at all as far as I
know.
The detection may work with WSL (this I did not test), however, it does
not work with git-bash (which is my usecase). Therefore forcing it to
"alacritty" seems reasonable.