Redefine "white" to a dark color on solarized_light.
Now, this is a bit controversial. After all "white" pretty much defines
a color. However, some applications (armclang in particular) use "white"
for highlighting text. This becomes unreadable on the solarized_light
background. The problem actually is with the application abusing the
color, redefining the color seems to be a feasible workaround (until a
proper application WANTS white text).