Disable mini.pairs.
After quite lengthy evaluation, the conclusion is that automatic
insertion of pairs is more often not desired. It's convenient sometimes,
but very often brackets/quotes are inserted after the conents of the
pair, in which case it just leads to too many pairs being created which
need to be deleted again.